change syntax

+++ Dangerfile
@@ -102,8 +102,9 @@ info = 'After the PR merged, attach the run result within the pipeline/jenkins'\
 # if official_reviewer.any? { |x| list_approval.include?(x) }
 github.api.add_label(repo, label, 'C05472') unless repo_label_name.include?(label)
 pr_comment_list.map do |u|
-                      github.api.delete_comment(repo, u['id'])
-                      if pr_comment_body.map(&:downcase).find { |e| /after the pr merged/ =~ e }
+  github.api.delete_comment(repo, u['id']) if pr_comment_body.map(&:downcase).find do |e|
+    /after the pr merged/ =~ e
+  end
 end
 
 if pr_comment_body.map(&:downcase).find { |e| /pr score/ =~ e }
